+++ CHANGELOG.md
@@ -16,6 +16,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Tree` and `TreeItem`
   - Updated docs and test suite
 
+## [0.9.4] - 2020-06-29
+
+### Added
+
+- `autoResize` added to `InputText` and `Select`, allowing component width to adjust to the current value
+
+### Removed
+
+- `InputText` props `prefix` and `suffix` are replaced with `before` and `after`, which accept `ReactNode`
+
+### Fixed
+
+- `InputTimeSelect` tab key behavior
+
+## [0.9.3] - 2020-06-26
+
+### Fixed
+
+- `MenuItem`
+  - properly renders when `[aria-current='false']`
+  - is now exported as Styled Component
+  - renders properly in Safari / older Chrome implementations
+
+## [0.9.2] - 2020-06-25
+
+### Added
+
+- `MenuList` now supports `disabled` property
+- `Popover` offset issue
+
+### Fixed
+
+- `MenuList` `groupDividers` now displayed correctly
+- Icon: `Pivot`
+
 ## [0.9.1]- 2020-06-24
 
 ### Changed